feat: SSIM状态机+中间帧选择+曝光调整
- 用三状态机(TRACKING/IDLE/COOLDOWN)替代旧SSIM导数U型判断
- 奇偶法维护跟踪队列，触发时取队列头部(即中间帧)送FPGA推理
- ArmControl: delay支持浮点数
- camera: 曝光7000us，去掉多余日志

+++ ArmControl/api.py
@@ -171,7 +171,7 @@ def grab_task(data: dict):
     logger.info("分拣开始, flags=%s", data.get('flags'))
 
     speed = int(data['speed'])  # 舵机运动速度 (ms)
-    delay = int(data.get('time', 0))  # 传送带延迟 (秒)
+    delay = float(data.get('time', 0))  # 传送带延迟 (秒)
     flags = data['flags']  # "OK" 或 "NG"
 
     # --- 步骤0: 等待铝片从相机位置行进到吸取点 ---
